JS
Added variables to contain;
Button colors, the game pattern, the users clicked pattern, and the level.

Added the function for the next sequence to generate random number and assign the value to the array.

Added event handler function for on click then push those values to the users array.

Added audio function and assigned to both user and random sequence function.

Added animation function for on press to add the CSS class created for the effect of a button press.

Added event handler for keypress.